backend/requirements_simple.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
face-recognition
mediapipe
pydantic-settings
python-dotenv
motor
requests
